+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21.1-4.0.20] - 2025-04-21
+### Fixed
+- Revert materials with empty tags not being loaded. In many cases this would cause most materials to erroneously fail to load on the first time loading a world.
+  - Instead, materials now have an `isValid` method which checks the tags. If the material is determined to be invalid, it will stay loaded but will not be displayed in most contexts.
+
 ## [1.21.1-4.0.19] - 2025-04-20
 ### Added
 - Additive property for materials. Additive materials can be used to make compound materials, but cannot be used to craft gear directly.